+++ app/controllers/roles_controller.rb
@@ -1,9 +1,30 @@
 # frozen_string_literal: true
 
+# We intentionally call @feature_flags.enabled? multiple times and don't
+# factor it out to make these checks easy to discover.
+# :reek:RepeatedConditional
 class RolesController < RestController
   include AuthorizeResource
 
+  ROLES_API_EXTENSION_KIND = :roles_api
+
   before_action :current_user
+
+  def initialize(
+    *args,
+    extension_repository: Conjur::Extension::Repository.new,
+    feature_flags: Rails.application.config.feature_flags,
+    **kwargs
+  )
+    super(*args, **kwargs)
+
+    @feature_flags = feature_flags
+
+    # If role API extensions are enabled, load the registered callbacks
+    @extensions = if @feature_flags.enabled?(:roles_api_extensions)
+      extension_repository.extension(kind: ROLES_API_EXTENSION_KIND)
+    end
+  end
 
   def show
     render(json: role.as_json.merge(members: role.memberships))
@@ -28,7 +49,7 @@ class RolesController < RestController
   #
   # For each membership, return the full details of the grant as a
   # JSON object.
-  # 
+  #
   #
   # +params[:filter]+ and +params[:count]+ are handled as for +#all_memberships+
   #
@@ -36,7 +57,7 @@ class RolesController < RestController
     memberships = filtered_roles(role.direct_memberships_dataset(filter_params), membership_filter)
     render_dataset(memberships)
   end
-  
+
   # Find all members of this role.
   #
   # For each member, return the full details of the grant as a
@@ -60,12 +81,20 @@ class RolesController < RestController
   #
   # This API endpoint exists to manage group entitlements through
   # the UI or other integrations outside of loading a policy.
+  #
+  # We intentionally call @feature_flags.enabled? multiple times and don't
+  # factor it out to make these checks easy to discover.
+  # :reek:DuplicateMethodCall
   def add_member
     authorize(:create, policy)
 
     member_id = params[:member]
     member = Role[member_id]
     raise Exceptions::RecordNotFound, member_id unless member
+
+    if @feature_flags.enabled?(:roles_api_extensions)
+      @extensions.call(:before_add_member, role: role, member: member)
+    end
 
     # If membership is already granted, grant_to will return nil.
     # In this case, don't emit an audit record.
@@ -80,6 +109,15 @@ class RolesController < RestController
       )
     end
 
+    if @feature_flags.enabled?(:roles_api_extensions)
+      @extensions.call(
+        :after_add_member,
+        role: role,
+        member: member,
+        membership: membership
+      )
+    end
+
     head(:no_content)
   end
 
@@ -87,12 +125,25 @@ class RolesController < RestController
   #
   # This API endpoint exists to manage group entitlements through
   # the UI or other integrations outside of loading a policy.
+  #
+  # We intentionally call @feature_flags.enabled? multiple times and don't
+  # factor it out to make these checks easy to discover.
+  # :reek:DuplicateMethodCall
   def delete_member
     authorize(:update, policy)
 
     member_id = params[:member]
     membership = role.memberships_dataset.where(member_id: member_id).first
     raise Exceptions::RecordNotFound, member_id unless membership
+
+    if @feature_flags.enabled?(:roles_api_extensions)
+      @extensions.call(
+        :before_delete_member,
+        role: role,
+        member: Role[member_id],
+        membership: membership
+      )
+    end
 
     membership.destroy
 
@@ -104,6 +155,15 @@ class RolesController < RestController
         client_ip: request.ip
       )
     )
+
+    if @feature_flags.enabled?(:roles_api_extensions)
+      @extensions.call(
+        :after_delete_member,
+        role: role,
+        member: Role[member_id],
+        membership: membership
+      )
+    end
 
     head(:no_content)
   end
@@ -132,7 +192,7 @@ class RolesController < RestController
   def filter_params
     request.query_parameters.slice(:search, :kind).symbolize_keys
   end
-  
+
   def render_params
     # Rails 5 requires parameters to be explicitly permitted before converting
     # to Hash.  See: https://stackoverflow.com/a/46029524
@@ -141,7 +201,7 @@ class RolesController < RestController
       .slice(*allowed_params).to_h.symbolize_keys
   end
 
-  def membership_filter        
+  def membership_filter
     filter = params[:filter]
     filter = Array(filter).map{ |id| Role.make_full_id(id, account) } if filter
     return filter
@@ -155,7 +215,7 @@ class RolesController < RestController
     resp = count_only?  ? count_payload(dataset) :
            block_given? ? yield(dataset)         : dataset.all
 
-    render(json: resp)    
+    render(json: resp)
   end
 
   def count_only?
@@ -165,6 +225,4 @@ class RolesController < RestController
   def count_payload(dataset)
     { count: dataset.count }
   end
-
-
 end